+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
-All notable changes to this project will be documented in this file.
+Drewl Meta Preview changelog
+
+* This project's changelog is formatted based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
+* This project follows [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
@@ -9,5 +12,9 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - LICENSE, CONTRIBUTING, CODE OF CONDUCT, README and CHANGELOG files
-- Minor CSS styles updates
 - Real-time update when changing inputs value on YOAST edit screen
+- Fallback when Yoast is not installed
+
+### Changed
+
+- Minor CSS style updates